+++ node/src/types/kernel/api/block/appendBlock/index.d.ts
@@ -17,12 +17,5 @@
 
 
 /* directories */
-export * as appendBlock from "./appendBlock";
-export * as getBlockKramdown from "./getBlockKramdown";
-export * as getChildBlocks from "./getChildBlocks";
-export * as insertBlock from "./insertBlock";
-export * as prependBlock from "./prependBlock";
 
 /* files */
-export * as getBlockBreadcrumb from "./getBlockBreadcrumb";
-export * as getDocInfo from "./getDocInfo";